Add files via upload
Added a Hand class. This class defines what the dealer and player will have for their cards. It goes over the kind of card, the value of each card, and a separate function for aces. Their is also another function for adding cards to the hand. These are all made so the game can function.