Add files via upload
This file contains a lot of the rules for the game. This includes prompts to take bets, showing cards to the player, prompt to choose an action, value that cannot be exceeded, value that will allow to add more cards, value the dealer cannot exceed (It makes it so he does not bust).